151673570 correct upgradekvm command
Read entries in kvm 'microgateway'.
For each of below entries call 'Update an entry in an environment-scoped KVM' if already present and
call 'Create an entry in an environment-scoped KVM' if not present.

Private_key_kid
Public_key1
Public_key1_kid

Make all 3 calls in parallel and print success message after all are finished.
Show error message if any of the calls fails.